Move example main to separate file
